clean up OnDemandPIDResolver to better isolate the linux vs non-linux impl split
